+++ index.d.ts
@@ -3126,6 +3126,11 @@ declare namespace nkruntime {
         status?: string
     }
 
+    export interface StorageIndexResult {
+      objects: StorageObject[]
+      cursor: string | null
+    }
+
     export interface EnvelopeStatusUpdate {
         statusUpdate: StatusUpdateMessage
     }
@@ -4990,7 +4995,7 @@ declare namespace nkruntime {
          * @returns A list of storage objects matching the query criteria.
          * @throws {TypeError, GoError}
          */
-        storageIndexList(indexName: string, query: string, limit: number, order?: string[], callerId?: string | void): StorageObject[];
+        storageIndexList(indexName: string, query: string, limit: number, order?: string[], callerId?: string | void, cursor?: string): StorageIndexResult;
 
         /**
          * Get Satori object.